+++ src/Controls/tests/UITests/Tests/CollectionView/Selection/CollectionViewUITests.PreselectedItem.cs
@@ -1,0 +1,26 @@
+﻿using NUnit.Framework;
+using UITest.Appium;
+using UITest.Core;
+
+namespace Microsoft.Maui.AppiumTests
+{
+	[Category(TestCategory.CollectionView)]
+	public class CollectionViewPreselectedItemUITests : CollectionViewUITests
+	{
+		public CollectionViewPreselectedItemUITests(TestDevice device)
+			: base(device)
+		{
+		}
+
+		[Test]
+		public void PreselectedItemCollectionView()
+		{
+			App.Click("PreselectedItem");
+			App.WaitForElement("TestCollectionView");
+
+			// 1. Check the preselected item.
+			App.WaitForElement("WaitForStubControl");
+			VerifyScreenshot();
+		}
+	}
+}
