+++ src/Controls/tests/UITests/Tests/CollectionView/Selection/CollectionViewUITests.PreselectedItems.cs
@@ -1,0 +1,26 @@
+﻿using NUnit.Framework;
+using UITest.Appium;
+using UITest.Core;
+
+namespace Microsoft.Maui.AppiumTests
+{
+	[Category(TestCategory.CollectionView)]
+	public class CollectionViewPreselectedItemsUITests : CollectionViewUITests
+	{
+		public CollectionViewPreselectedItemsUITests(TestDevice device)
+			: base(device)
+		{
+		}
+
+		[Test]
+		public void PreselectedItemsCollectionView()
+		{
+			App.Click("PreselectedItems");
+			App.WaitForElement("TestCollectionView");
+
+			// 1. Check the preselected items.
+			App.WaitForElement("WaitForStubControl");
+			VerifyScreenshot();
+		}
+	}
+}
