+++ src/TestUtils/src/UITest.Appium/AppiumIOSApp.cs
@@ -23,9 +23,9 @@ namespace UITest.Appium
 				{
 					var appId = Config.GetProperty<string>("AppId") ?? throw new InvalidOperationException($"{nameof(AppState)} could not get the appid property");
 					var state = _driver?.ExecuteScript("mobile: queryAppState", new Dictionary<string, object>
-						{
-							{ "bundleId", appId },
-						});
+					{
+						{ "bundleId", appId },
+					});
 
 					// https://developer.apple.com/documentation/xctest/xcuiapplicationstate?language=objc
 					return Convert.ToInt32(state) switch
